LPS-86003 Rename, type: git ls-files **StrangersNotAllowedException**
