+++ modules/apps/login/login-authentication-opensso-web/src/main/java/com/liferay/login/authentication/opensso/web/internal/servlet/taglib/OpenSSOBottomJSPDynamicInclude.java
@@ -28,10 +28,10 @@ import com.liferay.portal.kernel.settings.CompanyServiceSettingsLocator;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.security.sso.opensso.StrangersNotAllowedException;
 import com.liferay.portal.security.sso.opensso.configuration.OpenSSOConfiguration;
 import com.liferay.portal.security.sso.opensso.constants.OpenSSOConstants;
 import com.liferay.portal.security.sso.opensso.constants.OpenSSOWebKeys;
+import com.liferay.portal.security.sso.opensso.exception.StrangersNotAllowedException;
 
 import java.io.IOException;
