+++ modules/apps/portal-security-sso/portal-security-sso-opensso-api/bnd.bnd
@@ -2,6 +2,6 @@ Bundle-Name: Liferay Portal Security SSO OpenSSO API
 Bundle-SymbolicName: com.liferay.portal.security.sso.opensso.api
 Bundle-Version: 4.0.0
 Export-Package:\
-	com.liferay.portal.security.sso.opensso,\
 	com.liferay.portal.security.sso.opensso.configuration,\
-	com.liferay.portal.security.sso.opensso.constants
+	com.liferay.portal.security.sso.opensso.constants,\
+	com.liferay.portal.security.sso.opensso.exception
