+++ modules/apps/portal-security-sso/portal-security-sso-opensso-api/src/main/java/com/liferay/portal/security/sso/opensso/exception/StrangersNotAllowedException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.opensso;
+package com.liferay.portal.security.sso.opensso.exception;
 
 import com.liferay.portal.kernel.exception.PortalException;
