+++ modules/apps/portal-security-sso/portal-security-sso-opensso-api/src/main/resources/com/liferay/portal/security/sso/opensso/exception/packageinfo
