+++ modules/apps/portal-security-sso/portal-security-sso-opensso-impl/src/main/java/com/liferay/portal/security/sso/opensso/internal/auto/login/OpenSSOAutoLogin.java
@@ -43,10 +43,10 @@ import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.security.exportimport.UserImporter;
-import com.liferay.portal.security.sso.opensso.StrangersNotAllowedException;
 import com.liferay.portal.security.sso.opensso.configuration.OpenSSOConfiguration;
 import com.liferay.portal.security.sso.opensso.constants.OpenSSOConstants;
 import com.liferay.portal.security.sso.opensso.constants.OpenSSOWebKeys;
+import com.liferay.portal.security.sso.opensso.exception.StrangersNotAllowedException;
 import com.liferay.portal.util.PropsValues;
 
 import java.util.Calendar;
